🚨 Changed linter configuration (max-line-length=119).

+++ plana/apps/users/tests/views/test_views_association_user.py
@@ -57,9 +57,7 @@ class AssociationUserViewsTests(TestCase):
             "username": cls.president_user_name,
             "password": "motdepasse",
         }
-        cls.response_president = cls.president_student_client.post(
-            url_login, data_president
-        )
+        cls.response_president = cls.president_student_client.post(url_login, data_president)
 
         # Start a manager client used in some tests
         cls.manager_general_user_id = 3
@@ -100,9 +98,7 @@ class AssociationUserViewsTests(TestCase):
         response_student = self.student_client.get("/users/associations/")
         self.assertEqual(response_student.status_code, status.HTTP_200_OK)
 
-        associations_user_cnt = AssociationUser.objects.filter(
-            user_id=self.student_user_id
-        ).count()
+        associations_user_cnt = AssociationUser.objects.filter(user_id=self.student_user_id).count()
         content = json.loads(response_student.content.decode("utf-8"))
         self.assertEqual(len(content), associations_user_cnt)
 
@@ -113,9 +109,7 @@ class AssociationUserViewsTests(TestCase):
         - A student user can execute this request.
         - A student user gets correct association user list data based on search filters.
         """
-        response_president = self.president_student_client.get(
-            "/users/associations/?association_id=2"
-        )
+        response_president = self.president_student_client.get("/users/associations/?association_id=2")
         associations_user_cnt = AssociationUser.objects.filter(association_id=2).count()
         content = json.loads(response_president.content.decode("utf-8"))
         self.assertEqual(len(content), associations_user_cnt)
@@ -142,32 +136,18 @@ class AssociationUserViewsTests(TestCase):
         - Filter by institutions is possible.
         """
 
-        associations_user_validated_cnt = AssociationUser.objects.filter(
-            is_validated_by_admin=False
-        ).count()
-        response_validated_asso = self.manager_client.get(
-            "/users/associations/?is_validated_by_admin=false"
-        )
-        content_validated_asso = json.loads(
-            response_validated_asso.content.decode("utf-8")
-        )
+        associations_user_validated_cnt = AssociationUser.objects.filter(is_validated_by_admin=False).count()
+        response_validated_asso = self.manager_client.get("/users/associations/?is_validated_by_admin=false")
+        content_validated_asso = json.loads(response_validated_asso.content.decode("utf-8"))
         self.assertEqual(len(content_validated_asso), associations_user_validated_cnt)
 
         institutions_ids = [2, 3]
         associations_user_institutions_cnt = AssociationUser.objects.filter(
-            association_id__in=Association.objects.filter(
-                institution_id__in=institutions_ids
-            ).values_list("id")
+            association_id__in=Association.objects.filter(institution_id__in=institutions_ids).values_list("id")
         ).count()
-        response_institutions_asso = self.manager_client.get(
-            "/users/associations/?institutions=2,3,"
-        )
-        content_institutions_asso = json.loads(
-            response_institutions_asso.content.decode("utf-8")
-        )
-        self.assertEqual(
-            len(content_institutions_asso), associations_user_institutions_cnt
-        )
+        response_institutions_asso = self.manager_client.get("/users/associations/?institutions=2,3,")
+        content_institutions_asso = json.loads(response_institutions_asso.content.decode("utf-8"))
+        self.assertEqual(len(content_institutions_asso), associations_user_institutions_cnt)
 
     def test_anonymous_get_association_user_detail(self):
         """
@@ -175,9 +155,7 @@ class AssociationUserViewsTests(TestCase):
 
         - An anonymous user cannot execute this request.
         """
-        response_anonymous = self.anonymous_client.get(
-            f"/users/{self.unvalidated_user_id}/associations/"
-        )
+        response_anonymous = self.anonymous_client.get(f"/users/{self.unvalidated_user_id}/associations/")
         self.assertEqual(response_anonymous.status_code, status.HTTP_401_UNAUTHORIZED)
 
     def test_manager_get_unexisting_association_user(self):
@@ -195,9 +173,7 @@ class AssociationUserViewsTests(TestCase):
 
         - A student user cannot execute this request.
         """
-        response_student = self.student_client.get(
-            f"/users/{self.unvalidated_user_id}/associations/"
-        )
+        response_student = self.student_client.get(f"/users/{self.unvalidated_user_id}/associations/")
         self.assertEqual(response_student.status_code, status.HTTP_403_FORBIDDEN)
 
     def test_manager_get_association_user_detail(self):
@@ -207,15 +183,11 @@ class AssociationUserViewsTests(TestCase):
         - A manager user can execute this request.
         - Correct data is returned.
         """
-        response_manager = self.manager_client.get(
-            f"/users/{self.student_user_id}/associations/"
-        )
+        response_manager = self.manager_client.get(f"/users/{self.student_user_id}/associations/")
         self.assertEqual(response_manager.status_code, status.HTTP_200_OK)
 
         user_associations = AssociationUser.objects.filter(user_id=self.student_user_id)
-        user_associations_requested = json.loads(
-            response_manager.content.decode("utf-8")
-        )
+        user_associations_requested = json.loads(response_manager.content.decode("utf-8"))
         self.assertEqual(len(user_associations_requested), len(user_associations))
 
     def test_get_link_association_user(self):
@@ -225,9 +197,7 @@ class AssociationUserViewsTests(TestCase):
         - Always returns a 405 no matter which role is trying to access it.
         """
         response_manager = self.manager_client.get("/users/999/associations/999")
-        self.assertEqual(
-            response_manager.status_code, status.HTTP_405_METHOD_NOT_ALLOWED
-        )
+        self.assertEqual(response_manager.status_code, status.HTTP_405_METHOD_NOT_ALLOWED)
 
     def test_anonymous_post_association_user_404_user(self):
         """
@@ -330,9 +300,7 @@ class AssociationUserViewsTests(TestCase):
 
         - Link cannot be added to a user without a group where associations can be linked.
         """
-        GroupInstitutionFundUser.objects.filter(
-            user_id=self.unvalidated_user_id, group_id=5
-        ).delete()
+        GroupInstitutionFundUser.objects.filter(user_id=self.unvalidated_user_id, group_id=5).delete()
         response_anonymous = self.anonymous_client.post(
             "/users/associations/",
             {
@@ -420,9 +388,7 @@ class AssociationUserViewsTests(TestCase):
         )
         self.assertTrue(len(mail.outbox))
         self.assertEqual(response_student.status_code, status.HTTP_201_CREATED)
-        user_asso = AssociationUser.objects.get(
-            user_id=self.student_user_id, association_id=5
-        )
+        user_asso = AssociationUser.objects.get(user_id=self.student_user_id, association_id=5)
         self.assertFalse(user_asso.is_validated_by_admin)
 
     def test_manager_misc_post_association_user_forbidden(self):
@@ -486,12 +452,8 @@ class AssociationUserViewsTests(TestCase):
 
         - Always returns a 405 no matter which role is trying to access it.
         """
-        response_manager = self.manager_client.put(
-            "/users/999/associations/999", {"is_treasurer": True}
-        )
-        self.assertEqual(
-            response_manager.status_code, status.HTTP_405_METHOD_NOT_ALLOWED
-        )
+        response_manager = self.manager_client.put("/users/999/associations/999", {"is_treasurer": True})
+        self.assertEqual(response_manager.status_code, status.HTTP_405_METHOD_NOT_ALLOWED)
 
     def test_anonymous_patch_association_user(self):
         """
@@ -499,9 +461,7 @@ class AssociationUserViewsTests(TestCase):
 
         - An anonymous user cannot execute this request.
         """
-        response_anonymous = self.anonymous_client.patch(
-            f"/users/{self.unvalidated_user_id}/associations/2"
-        )
+        response_anonymous = self.anonymous_client.patch(f"/users/{self.unvalidated_user_id}/associations/2")
         self.assertEqual(response_anonymous.status_code, status.HTTP_401_UNAUTHORIZED)
 
     def test_student_patch_association_user(self):
@@ -522,9 +482,7 @@ class AssociationUserViewsTests(TestCase):
 
         - A student president cannot remove his own privileges.
         """
-        asso_user = AssociationUser.objects.get(
-            user_id=self.president_user_id, is_president=True
-        )
+        asso_user = AssociationUser.objects.get(user_id=self.president_user_id, is_president=True)
         response_president = self.president_student_client.patch(
             f"/users/{self.president_user_id}/associations/{asso_user.association_id}",
             {"is_president": False},
@@ -539,17 +497,13 @@ class AssociationUserViewsTests(TestCase):
         - A student president of an association cannot change the validation status.
         """
         association_id = 2
-        AssociationUser.objects.get(
-            user_id=self.student_user_id, association_id=association_id
-        )
+        AssociationUser.objects.get(user_id=self.student_user_id, association_id=association_id)
         response_president = self.president_student_client.patch(
             f"/users/{self.student_user_id}/associations/{association_id}",
             {"is_validated_by_admin": False},
             content_type="application/json",
         )
-        AssociationUser.objects.get(
-            user_id=self.student_user_id, association_id=association_id
-        )
+        AssociationUser.objects.get(user_id=self.student_user_id, association_id=association_id)
         self.assertEqual(response_president.status_code, status.HTTP_403_FORBIDDEN)
 
     def test_student_patch_association_user_president_forbidden(self):
@@ -568,9 +522,7 @@ class AssociationUserViewsTests(TestCase):
             },
             content_type="application/json",
         )
-        asso_user = AssociationUser.objects.get(
-            user_id=self.student_user_id, association_id=association_id
-        )
+        asso_user = AssociationUser.objects.get(user_id=self.student_user_id, association_id=association_id)
         self.assertEqual(response_president.status_code, status.HTTP_403_FORBIDDEN)
         self.assertFalse(asso_user.is_president)
 
@@ -611,9 +563,7 @@ class AssociationUserViewsTests(TestCase):
             },
             content_type="application/json",
         )
-        asso_user = AssociationUser.objects.get(
-            user_id=self.student_user_id, association_id=association_id
-        )
+        asso_user = AssociationUser.objects.get(user_id=self.student_user_id, association_id=association_id)
         self.assertEqual(response_president.status_code, status.HTTP_200_OK)
         self.assertTrue(len(mail.outbox))
         self.assertEqual(asso_user.can_be_president_from, datetime.date(2023, 3, 22))
@@ -630,9 +580,7 @@ class AssociationUserViewsTests(TestCase):
             },
             content_type="application/json",
         )
-        asso_user = AssociationUser.objects.get(
-            user_id=self.student_user_id, association_id=association_id
-        )
+        asso_user = AssociationUser.objects.get(user_id=self.student_user_id, association_id=association_id)
         self.assertEqual(response_president.status_code, status.HTTP_200_OK)
         self.assertEqual(asso_user.can_be_president_to, None)
         self.assertFalse(asso_user.is_president)
@@ -648,9 +596,7 @@ class AssociationUserViewsTests(TestCase):
             },
             content_type="application/json",
         )
-        asso_user = AssociationUser.objects.get(
-            user_id=self.student_user_id, association_id=association_id
-        )
+        asso_user = AssociationUser.objects.get(user_id=self.student_user_id, association_id=association_id)
         self.assertEqual(response_president.status_code, status.HTTP_200_OK)
         self.assertFalse(asso_user.is_president)
         self.assertFalse(asso_user.is_secretary)
@@ -677,9 +623,7 @@ class AssociationUserViewsTests(TestCase):
         - A manager can execute this request.
         - Serializer fields must be valid.
         """
-        asso_user = AssociationUser.objects.get(
-            user_id=self.president_user_id, is_president=True
-        )
+        asso_user = AssociationUser.objects.get(user_id=self.president_user_id, is_president=True)
         response = self.manager_client.patch(
             f"/users/{self.president_user_id}/associations/{asso_user.association_id}",
             {"is_president": "bad format"},
@@ -699,23 +643,17 @@ class AssociationUserViewsTests(TestCase):
         - A manager can add a president to an association without one.
         """
         association_id = 2
-        asso_user = AssociationUser.objects.get(
-            user_id=self.student_user_id, association_id=association_id
-        )
+        asso_user = AssociationUser.objects.get(user_id=self.student_user_id, association_id=association_id)
         response = self.manager_client.patch(
             f"/users/{self.student_user_id}/associations/{association_id}",
             {"is_president": True},
             content_type="application/json",
         )
-        asso_user = AssociationUser.objects.get(
-            user_id=self.student_user_id, association_id=association_id
-        )
+        asso_user = AssociationUser.objects.get(user_id=self.student_user_id, association_id=association_id)
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertTrue(asso_user.is_president)
 
-        old_president = AssociationUser.objects.get(
-            user_id=self.president_user_id, association_id=association_id
-        )
+        old_president = AssociationUser.objects.get(user_id=self.president_user_id, association_id=association_id)
         self.assertFalse(old_president.is_president)
 
         self.assertFalse(len(mail.outbox))
@@ -724,9 +662,7 @@ class AssociationUserViewsTests(TestCase):
             {"is_validated_by_admin": True},
             content_type="application/json",
         )
-        asso_user = AssociationUser.objects.get(
-            user_id=self.unvalidated_user_id, association_id=association_id
-        )
+        asso_user = AssociationUser.objects.get(user_id=self.unvalidated_user_id, association_id=association_id)
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertTrue(asso_user.is_validated_by_admin)
         self.assertTrue(len(mail.outbox))
@@ -744,9 +680,7 @@ class AssociationUserViewsTests(TestCase):
             {"is_president": True},
             content_type="application/json",
         )
-        asso_user = AssociationUser.objects.get(
-            user_id=self.student_user_id, association_id=association_id
-        )
+        asso_user = AssociationUser.objects.get(user_id=self.student_user_id, association_id=association_id)
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertTrue(asso_user.is_president)
 
@@ -757,9 +691,7 @@ class AssociationUserViewsTests(TestCase):
         - A manager can execute this request.
         - Link between member and association is correctly updated.
         """
-        asso_user = AssociationUser.objects.get(
-            user_id=self.president_user_id, is_president=True
-        )
+        asso_user = AssociationUser.objects.get(user_id=self.president_user_id, is_president=True)
         response = self.manager_client.patch(
             f"/users/{self.president_user_id}/associations/{asso_user.association_id}",
             {"is_president": False},
@@ -814,9 +746,7 @@ class AssociationUserViewsTests(TestCase):
         - A student user cannot execute this request.
         """
         asso_user = AssociationUser.objects.get(user_id=self.unvalidated_user_id)
-        response_student = self.student_client.delete(
-            f"/users/{self.unvalidated_user_id}/associations/{asso_user.id}"
-        )
+        response_student = self.student_client.delete(f"/users/{self.unvalidated_user_id}/associations/{asso_user.id}")
         self.assertEqual(response_student.status_code, status.HTTP_403_FORBIDDEN)
 
     def test_manager_delete_association_user_404_user(self):
@@ -825,14 +755,10 @@ class AssociationUserViewsTests(TestCase):
 
         - Cannot remove link to an association from a non-existing user.
         """
-        response = self.manager_client.get(
-            f"/users/{self.student_user_id}/associations/"
-        )
+        response = self.manager_client.get(f"/users/{self.student_user_id}/associations/")
         first_user_association_id = response.data[0]["id"]
 
-        response_delete = self.manager_client.delete(
-            f"/users/9999/associations/{str(first_user_association_id)}"
-        )
+        response_delete = self.manager_client.delete(f"/users/9999/associations/{str(first_user_association_id)}")
         self.assertEqual(response_delete.status_code, status.HTTP_404_NOT_FOUND)
 
     def test_manager_delete_association_user_404_association(self):
@@ -841,9 +767,7 @@ class AssociationUserViewsTests(TestCase):
 
         - Cannot remove link from a non-existing association to a user.
         """
-        response_delete = self.manager_client.delete(
-            f"/users/{self.student_user_id}/associations/99"
-        )
+        response_delete = self.manager_client.delete(f"/users/{self.student_user_id}/associations/99")
         self.assertEqual(response_delete.status_code, status.HTTP_404_NOT_FOUND)
 
     def test_manager_delete_association_user(self):
@@ -855,9 +779,7 @@ class AssociationUserViewsTests(TestCase):
         - A manager user can execute this request.
         - The link between an association and a user is deleted.
         """
-        response = self.manager_client.get(
-            f"/users/{self.student_user_id}/associations/"
-        )
+        response = self.manager_client.get(f"/users/{self.student_user_id}/associations/")
         first_user_association_id = response.data[0]["id"]
 
         response_delete = self.manager_client.delete(
@@ -865,6 +787,4 @@ class AssociationUserViewsTests(TestCase):
         )
         self.assertEqual(response_delete.status_code, status.HTTP_204_NO_CONTENT)
         with self.assertRaises(ObjectDoesNotExist):
-            AssociationUser.objects.get(
-                user_id=self.student_user_id, association_id=first_user_association_id
-            )
+            AssociationUser.objects.get(user_id=self.student_user_id, association_id=first_user_association_id)
